requirements.txt
psychopy>=2021.2.0
mne>=0.24.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
pyserial>=3.5